refactor(input): ♻️ refactored import and exports
