+++ src/components/Input/view/Input.tsx
@@ -5,7 +5,7 @@ import { InputProps } from './base/types';
 import { PropsWithoutChildren } from '../../../types';
 import { InputFixProps } from './composed/withFix/types';
 import { InputWrapper as InputWrapperWithFix } from './composed';
-import { StyledInputSpan } from './base/style';
+import { StyledInputSpan } from './composed/withFix/style/Input.style';
 
 export type InputType = FC<PropsWithoutChildren<InputFixProps & InputProps>>;
