+++ src/components/Input/view/base/style/parts/fillWidth/index.ts
@@ -1,0 +1,1 @@
+export * from './fullWidth';
