+++ src/components/Input/view/base/style/parts/fixPadding/index.ts
@@ -1,0 +1,1 @@
+export * from './fixPadding';
