+++ src/components/Input/view/base/style/parts/isDisabled/index.ts
@@ -1,0 +1,1 @@
+export * from './isDisabled';
