+++ src/components/Input/view/base/style/parts/isRound/index.ts
@@ -1,0 +1,1 @@
+export * from './isRound';
