+++ src/components/Input/view/base/style/parts/isRound/isRound.ts
@@ -1,6 +1,6 @@
 import { css } from 'styled-components';
-import { InputProps } from '../types';
-import { ThemeType } from '../../../../../theme';
+import { InputProps } from '../../../types';
+import { ThemeType } from '../../../../../../../theme';
 
 type IsRoundInput = (
     isRound: InputProps['isRound'],
