+++ src/components/Input/view/base/style/parts/size/index.ts
@@ -1,0 +1,1 @@
+export * from './size';
