+++ src/components/Input/view/base/style/parts/sizeDiff/index.ts
@@ -1,0 +1,1 @@
+export * from './sizeDiff';
